break up serverside event handlers into multiple files
