+++ server/eventHandlers/index.js
@@ -1,0 +1,19 @@
+const cursorMove = require('./cursorMove');
+const updateSelectedColor = require('./updateSelectedColor');
+const updateSelectedFrame = require('./updateSelectedFrame');
+const addNewFrame = require('./addNewFrame');
+const selectLayer = require('./selectLayer');
+const addNewLayer = require('./addNewLayer');
+const setPreviewLayer = require('./setPreviewLayer');
+const disconnect = require('./disconnect');
+
+module.exports = {
+  cursorMove,
+  updateSelectedColor,
+  updateSelectedFrame,
+  addNewFrame,
+  selectLayer,
+  addNewLayer,
+  setPreviewLayer,
+  disconnect
+};
