+++ server/index.js
@@ -4,16 +4,9 @@ const socketio = require('socket.io');
 const morgan = require('morgan');
 const chalk = require('chalk');
 const constants = require('../shared/constants');
-const {
-  initializeEmptySprite,
-  userFactory,
-  frameFactory,
-  layerFactory
-} = require('../shared/factories');
 
 const PORT = process.env.PORT || 3000;
-const loadData = require('./db/loadData');
-const saveData = require('./db/saveData');
+const pixalive = require('./pixalive.js');
 
 // initialize express
 const app = express();
@@ -60,152 +53,5 @@ const namespacedIo = io.of(/.*/).on(constants.MSG.CONNECT, socket => {
   });
 });
 
-/**************************************
- *   vv THE FUN STUFF STARTS HERE vv
- **************************************/
-
-// root of our server-side state tree
-// a hash of namespaces/sprites
-const state = {};
-
-namespacedIo.on(constants.MSG.CONNECT, async socket => {
-  // store our sprite hash and socket id
-  const spriteHash = socket.nsp.name.slice(1);
-  const socketId = socket.id.slice(socket.nsp.name.length + 1);
-
-  // does this namespace exist? if not, create it
-  if (!state[spriteHash]) {
-    try {
-      state[spriteHash] = await loadData(spriteHash);
-    } catch (err) {
-      console.error(err);
-    }
-  }
-
-  // make a new user object and add it
-  state[spriteHash].users[socketId] = userFactory(socketId);
-
-  // send the current drawing object
-  socket.emit(constants.MSG.SEND_SPRITE, state[spriteHash]);
-
-  // when a cursor moves...
-  socket.on(constants.MSG.CURSOR_MOVE, coords => {
-    // update their coords
-    state[spriteHash].users[socketId].x = coords.x;
-    state[spriteHash].users[socketId].y = coords.y;
-
-    // send only the cursor update
-    namespacedIo.emit(constants.MSG.CURSOR_UPDATE, {
-      x: coords.x,
-      y: coords.y,
-      socketId
-    });
-  });
-
-  //handle color update on user
-  socket.on(constants.MSG.UPDATE_SELECTED_COLOR, selectedColor => {
-    state[spriteHash].users[socketId].selectedColor = selectedColor;
-
-    namespacedIo.emit(constants.MSG.SEND_SPRITE, state[spriteHash]);
-  });
-
-  //handle selected frame
-  socket.on(constants.MSG.UPDATE_SELECTED_FRAME, selectedFrame => {
-    state[spriteHash].users[socketId].selectedFrame = selectedFrame;
-
-    namespacedIo.emit(constants.MSG.SEND_SPRITE, state[spriteHash]);
-  });
-
-  //add new frame
-  socket.on(constants.MSG.ADD_NEW_FRAME, () => {
-    //get new frameOrder value. People can delete a frame so
-    //new frame needs to be higher than current max frame
-    const arrayOfFrameKeys = state[spriteHash].frames.map(
-      frame => frame.frameOrder
-    );
-    const currentMax = Math.max(...arrayOfFrameKeys);
-    const newFrameOrder = currentMax + 1;
-
-    //make a new frame and add to frames
-    const newFrame = frameFactory(newFrameOrder);
-
-    // add new layers to single frame
-    const sampleFrame = state[spriteHash].frames[0];
-    sampleFrame.layers.forEach(layer => {
-      const h = layer.pixels.length;
-      const w = layer.pixels[0].length;
-      const newLayer = layerFactory(w, h, layer.layerOrder, layer.name);
-      newFrame.layers.push(newLayer);
-    });
-
-    // push in new frame
-    state[spriteHash].frames.push(newFrame);
-
-    //send updated sprite
-    namespacedIo.emit(constants.MSG.SEND_SPRITE, state[spriteHash]);
-  });
-
-   //select layer
-   socket.on(constants.MSG.SELECT_LAYER, selectedLayer => {
-    state[spriteHash].users[socketId].selectedLayer = selectedLayer;
-
-    namespacedIo.emit(constants.MSG.SEND_SPRITE, state[spriteHash]);
-  });
-
-  //add new layer to all frames
-  socket.on(constants.MSG.ADD_NEW_LAYER, () => {
-    //get new frameOrder value. People can delete a frame so
-    //new frame needs to be higher than current max frame
-    const selectedFrame = state[spriteHash].users[socketId].selectedFrame;
-    const arrayOfLayerKeys = state[spriteHash].frames[selectedFrame].layers.map(
-      layer => layer.layerOrder
-    );
-    const currentMax = Math.max(...arrayOfLayerKeys);
-    const newLayerOrder = currentMax + 1;
-
-    //make a new layer and add to frames
-    const w = state[spriteHash].frames[selectedFrame].layers[0].pixels[0].length
-    const h = state[spriteHash].frames[selectedFrame].layers[0].pixels.length
-
-
-    // add new layer to all frames
-
-    state[spriteHash].frames.forEach(frame => {
-      const newLayer = layerFactory(w, h, newLayerOrder);
-      frame.layers.push(newLayer)
-    })
-
-    //send updated sprite
-    namespacedIo.emit(constants.MSG.SEND_SPRITE, state[spriteHash]);
-  });
-
-  //set preview to true or false
-  socket.on(constants.MSG.SET_PREVIEW_LAYER, toggle => {
-
-    state[spriteHash].users[socketId].preview = toggle;
-
-    //send updated sprite
-    namespacedIo.emit(constants.MSG.SEND_SPRITE, state[spriteHash]);
-  })
-
-  // when this client leaves
-  socket.on(constants.MSG.DISCONNECT, async socket => {
-    // take the user out of the namespace/sprite
-    delete state[spriteHash].users[socketId];
-
-    // emit it so it updates for everyone still connectd
-    namespacedIo.emit(constants.MSG.SEND_SPRITE, state[spriteHash]);
-
-    // get the number of users left in this namespace
-    const usersLeft = Object.keys(state[spriteHash].users).length;
-
-    // if nobody left, free up the memory
-    if (!usersLeft) {
-      await saveData(state[spriteHash]);
-      console.log(
-        chalk.red(`index.js -> DELETING SPRITE -> spriteHash: ${spriteHash}`)
-      );
-      delete state[spriteHash];
-    }
-  });
-});
+// initialize the app
+pixalive(namespacedIo);
